requirements.txt
flask
openai
python-dotenv
gunicorn
speechrecognition
soundfile
pydub
requests
Pillow
langchain-core
langchain-google-genai
google-generativeai